+++ gmaven/src/main/kotlin/ru/rzn/gmyasoedov/gmaven/wizard/MavenNewProjectWizard.kt
@@ -24,14 +24,13 @@ import com.intellij.ui.dsl.builder.Panel
 import com.intellij.ui.dsl.builder.TopGap
 import com.intellij.ui.dsl.builder.bindSelected
 import ru.rzn.gmyasoedov.gmaven.GMavenConstants
-import ru.rzn.gmyasoedov.gmaven.GMavenConstants.GMAVEN
 import ru.rzn.gmyasoedov.gmaven.settings.MavenSettings
 import ru.rzn.gmyasoedov.maven.plugin.reader.model.SimpleMavenId
 import java.nio.file.Path
 
 class MavenNewProjectWizard : BuildSystemJavaNewProjectWizard {
 
-    override val name = GMAVEN
+    override val name = GMavenConstants.SYSTEM_ID.readableName
 
     override fun createStep(parent: JavaNewProjectWizard.Step) = Step(parent).chain(::AssetsStep)