+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ## [223.31] - 2025-01-07
 ### Added
 
+- rename to Easy Maven ([issue](https://github.com/grisha9/gmaven-plugin/issues/10))
 - move features from master
 ### Fixed
 - Split module - test scope